requirements.txt
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
django-import-export=3.1.0
drf-yasg==1.21.4
pytest==7.2.1
